feat: Songsanity (#111)
* Include first overrides of songsanity.

TODO: Adjust edge cases for 4 songs.

* Fix song of soaring still giving item.

Text still appears but SoS is not given.

TODO: NWBN soft locking.

* fix: Ensure we do not fully reset during moon crash, but only cycle flags.

* Adjust some stored text ID issues.

This ensures that song of healing and song text boxes work with vanilla as well.

TODO: Adjust HMS and Kaepora to check for ext data instead of inventory.

* Add kaepora ext data check.

* Add in HMS ext data check.

* Move NWBN to new patches.

Fix up and get NWBN working.

* Include song of time checks.

* Rename SoH function to be more inline with SoH and songs.

Decomp message context and place in gctx.

* Finalize sonata hooks.

* Include epona checks.

* Include get item override text for oath to order.

* Include song of storms override.

* Adjust input to include circle pad.

Update graphics to not flush process data cache when on emulator.

* Include progressive lullaby and new checks for lullaby.